Create README.md
This Roller Coaster project is meant to create several data visualizations that will give insight into the world of roller coasters.
Given two dataframes, one for roller coasters made of wood, and another for roller coasters made of steel, which are the normal categories they are split into. The main goal
is to write python functions that plot the rankings of given roller coasters over time as a line. This project uses line graphs, histograms, bar charts, pie charts, as well as a scatter plot to 
accurately describe the data in both dataframes. The dataframe has limited information about the rollercoasters, so further statistics could not be calculated, such as popularity of coasters 
based on seating types, height, speed, or length of coaster. Overall, this was an interesting project, as I am personally a big coaster fan. I may come back and find more data about the 
compaines that build these rollercoasters and determine which are the fastest, tallest, longest, or even the curviest of them all. I often wonder if there is a coaster that causes
the most patrons to vomit. Just my wonderings.